cmd/enter: Set failing command as active in main parser
This commit fixes handling of the following two cases:

  1. pebble enter subcommand --help
  2. pebble enter help subcommand

Help messages are handled in main.go when the executed command returns
flags.Error{Type:flags.ErrHelp}. When we pass the subcommand's return
value back to the main parser, the error handling there will use the
currently active command to print the help message. The active command
is enter but we want to print the subcommands help message.

Whenever the subcommand returns an error value, set it as active command
in the main parser, so that there error handling there works with the
failing subcommand.